pusssss

+++ example-app/app/Http/Controllers/MahasiswaController.php
@@ -14,7 +14,12 @@ class MahasiswaController extends Controller
      */
     public function index()
     {
+        if(auth()->user()->role == 'D'){
+            $mahasisw = Mahasiswa::where('user_id',auth()->user()->id)->get();
+            //select * from mahasiswas where user_id =1
+        }else{
         $mahasiswa = Mahasiswa::all();
+        }
         return view('mahasiswa.index')
                 ->with('mahasiswa', $mahasiswa);
 
@@ -126,6 +131,7 @@ class MahasiswaController extends Controller
      */
     public function destroy(Mahasiswa $mahasiswa)
     {
+        if(auth()->user()->cannot('update',$mahasiswa)){abort(403)};
         //dd($mahasiswa);
         File::delete('foto/'. $mahasiswa['url_foto']);
         $mahasiswa->delete(); // hapus data mahasiswa
